Refactor tool_shed basic_util into Galaxy.
